Add mastermind.rb file
This is a copy of tic-tac-toe.rb from a different project.  The Tic Tac
Toe specific details have been removed.  The TicTacToe class is renamed
to Mastermind.  Added constants and new private methods (empty right
now) for the Mastermind class.

Updated the GameManager class to use one player instead of two players.

The Tic Tac Toe game passed around the game_mark string ('X' or 'O'),
this is not relevant for Mastermind, so has been removed.

The API for Tic Tac Toe and GameManager are mostly the same, which made
code reuse suitable for this project.