Complete new_game method
The only thing that was missing in this method was creating the new
results array.  This will be an array of hashes, where each hash will
consist of the following:

  { :attempt <player code> :feedback <results> }

The feedback will be a string consisting of 'B', 'W', and/or '-'
characters.

This array will store the history of the player entries, which will end
up being printed to the screen as the game is played.