Convert attempt_code to a class method
This includes all of the attempt_code helper classes as well:

  process_exact_matches
  process_of_matches
  process_no_matches

The helper methods are set to private.

The attempt_code call in play? was updated as well.

The reason for making the attempt_code a class method is that a
computer AI will be implemented that will try to guess the secret code.
The algorithm that will be used will be the one created by Donald Knuth.
In order to implement this algorithm, several comparisons will need to
be made between codes and potential secret codes.  The attempt_code
method here will be used for that part.

The original object attempt_code method only compared a code with the
actual secret code created by the Mastermind instance.  Now it compares
any two codes and returns the feedback string for it.

The Donald Knuth algorithm is explained here:

  https://en.wikipedia.org/wiki/Mastermind_(board_game)

Under the "Worst case: Five-guess algorithm section.